Fixes bug
When we change the optimizer function it seems to zero all the stored data (like the mean averages and everything) which we don't want if it's the same optimizer (and ideally not at all). There's more weird stuff going on as I can't manage to replicate the same error as with use_clr_beta in some cases.